requirements.txt
fastapi
uvicorn
asyncpg
bcrypt
sqlalchemy
python-dotenv
pydantic-settings
python-jose